pyproject.toml
[project]
name = "ragtool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.9.6",
    "django>=5.2.1",
    "django-jazzmin>=3.0.1",
    "dj-database-url>=3.0.1",
    "django-cors-headers>=4.9.0",
    "openai>=2.24.0",
    "langchain>=0.3.26",
    "langchain-core>=0.3.69",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "mcp>=1.26.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "tavily-python>=0.7.12",
    "yfinance>=0.2.66",
    "sqlalchemy>=2.0.43",
    "aiosqlite>=0.21.0",
]
